+++ include/UECS/RandomAccessor.h
@@ -1,0 +1,12 @@
+#pragma once
+
+#include "CmptType.h"
+
+#include <set>
+
+namespace Ubpa::UECS {
+	class RandomAccessor {
+	public:
+		CmptAccessTypeSet types;
+	};
+}